feat: create Monster interface in a separate file and update references across the application
